+++ locallib.php
@@ -237,3 +237,52 @@ function theme_boost_campus_set_node_on_top(flat_navigation $flatnav, $nodename,
     // Return the modified changes.
     return $flatnav;
 }
+
+
+/**
+ * Provides the node for the in-course course settings.
+ *
+ * @return navigation_node.
+ */
+function theme_boost_campus_get_incourse_settings() {
+    global $COURSE, $PAGE;
+    // Initialize the node with false to prevent problems on pages that do not have a courseadmin node.
+    $node = false;
+    // If setting showsettingsincourse is enabled.
+    if (get_config('theme_boost_campus', 'showsettingsincourse') == 'yes') {
+        // Only search for the courseadmin node if we are within a course or a module context.
+        if ($PAGE->context->contextlevel == CONTEXT_COURSE || $PAGE->context->contextlevel == CONTEXT_MODULE) {
+            // Get the courseadmin node for the current page.
+            $node = $PAGE->settingsnav->find('courseadmin', navigation_node::TYPE_COURSE);
+            // If the setting 'incoursesettingsswitchtorole' is enabled add these to the $node.
+            if (get_config('theme_boost_campus', 'incoursesettingsswitchtorole') == 'yes' && !is_role_switched($COURSE->id)) {
+                // Build switch role link
+                // We could only access the existing menu item by creating the user menu and traversing it.
+                // So we decided to create this node from scratch with the values copied from Moodle core.
+                $roles = get_switchable_roles($PAGE->context);
+                if (is_array($roles) && (count($roles) > 0)) {
+                    // Define the properties for a new tab.
+                    $properties = array('text' => get_string('switchroleto', 'theme_boost_campus'),
+                                        'type' => navigation_node::TYPE_CONTAINER,
+                                        'key'  => 'switchroletotab');
+                    // Create the node.
+                    $switchroletabnode = new navigation_node($properties);
+                    // Add the tab to the course administration node.
+                    $node->add_node($switchroletabnode);
+                    // Add the available roles as children nodes to the tab content.
+                    foreach ($roles as $key => $role) {
+                        $properties = array('action' => new moodle_url('/course/switchrole.php',
+                            array('id'         => $COURSE->id,
+                                  'switchrole' => $key,
+                                  'returnurl'  => $PAGE->url->out_as_local_url(false),
+                                  'sesskey'    => sesskey())),
+                                            'type'   => navigation_node::TYPE_CUSTOM,
+                                            'text'   => $role);
+                        $switchroletabnode->add_node(new navigation_node($properties));
+                    }
+                }
+            }
+        }
+    }
+    return $node;
+}